idi inserting settings dialogs

+++ app/src/main/java/com/example/dirio/view/viewHolder/DiariesViewHolder.kt
@@ -1,12 +1,13 @@
 package com.example.dirio.view.viewHolder
 
-import android.view.View
-import android.view.View.OnLongClickListener
+import android.app.AlertDialog
+import android.widget.Toast
 import androidx.recyclerview.widget.RecyclerView
 import com.example.dirio.Repository.DailyEntitie
 import com.example.dirio.databinding.RowDailyBinding
 import com.example.dirio.listener.ListenerFragment
-import java.time.format.DateTimeFormatter
+
+
 
 class DiariesViewHolder(private val bind: RowDailyBinding, val listener: ListenerFragment) :
     RecyclerView.ViewHolder(bind.root) {
@@ -20,7 +21,17 @@ class DiariesViewHolder(private val bind: RowDailyBinding, val listener: Listene
         }
 
         bind.linerRowDaily.setOnLongClickListener {
-            listener.deleteDaiyl(daily)
+
+            val builder = AlertDialog.Builder(itemView.context)
+            builder.setMessage("Tem certeza que deseja deleter esta anotação?")
+                .setPositiveButton("Sim") { dialog, which ->
+                    listener.deleteDaiyl(daily)
+                }
+            builder.setNegativeButton("Não") { dialog, which ->
+                dialog.cancel()
+            }
+            builder.create().show()
+            Toast.makeText(itemView.context, "Pagina Deletada", Toast.LENGTH_SHORT).show()
             true
         }
